Move ShaderCursor implementation to a cpp file
